# Use `bitcode::Buffer` in benchmark, mirroring `fastbuf::Buffer`.